typing fix (corrected)

+++ balanced_clustering/utils/_emi.py
@@ -12,7 +12,7 @@ def _emi(a, b, R, C, N):
     # There are three major terms to the EMI equation, which are multiplied to
     # and then summed over varying nij values.
     # While nijs[0] will never be used, having it simplifies the indexing.
-    nijs = np.arange(0.0, float(max(np.max(a), np.max(b)) + 1), dtype=np.float64)
+    nijs = np.arange(0.0, float(max(np.max(a), np.max(b)) + 1))
     nijs[0] = 1  # Stops divide by zero warnings. As its not used, no issue.
     # term1 is nij / N
     term1 = nijs / N
